+++ project-base/tests/ShopBundle/Acceptance/acceptance/PageObject/Front/RegistrationPage.php
@@ -21,6 +21,7 @@ class RegistrationPage extends AbstractPage
         $this->tester->fillFieldByName('registration_form[email]', $email);
         $this->tester->fillFieldByName('registration_form[password][first]', $firstPassword);
         $this->tester->fillFieldByName('registration_form[password][second]', $secondPassword);
+        $this->tester->checkOptionByLabel('I agree with privacy policy');
         $this->tester->wait(TimedFormTypeExtension::MINIMUM_FORM_FILLING_SECONDS);
         $this->tester->clickByName('registration_form[save]');
     }